Move code to separate module
